Refactor StunMessage.hpp to add support for REQUESTED-TRANSPORT attribute
This commit refactors the StunMessage.hpp file to add support for the REQUESTED-TRANSPORT attribute. The REQUESTED-TRANSPORT attribute is used by the client to request a specific transport protocol for the allocated transport address. The commit introduces the ProtocolTransport enum, which specifies the desired protocol (TCP or UDP), and adds helper functions to convert the enum to a string and vice versa. Additionally, the commit adds the makeRequestTransport and parseRequestTransport functions to create and parse the REQUESTED-TRANSPORT attribute, respectively. The commit also includes some minor code cleanup and formatting changes.